Move memory_media_file_storage to the runtime package
We need to instantiate it on the fly when running in raw mode
