+++ lib/tests/streamlit/elements/video_test.py
@@ -20,7 +20,7 @@ import numpy as np
 
 import streamlit as st
 from streamlit.runtime.media_file_storage import MediaFileStorageError
-from streamlit.web.server.memory_media_file_storage import (
+from streamlit.runtime.memory_media_file_storage import (
     _calculate_file_id,
 )
 from streamlit.web.server.server import MEDIA_ENDPOINT
