+++ lib/tests/streamlit/runtime/runtime_test.py
@@ -34,8 +34,8 @@ from streamlit.runtime.runtime import (
     AsyncObjects,
 )
 from streamlit.runtime.uploaded_file_manager import UploadedFileRec
+from streamlit.runtime.memory_media_file_storage import MemoryMediaFileStorage
 from streamlit.watcher import event_based_path_watcher
-from streamlit.web.server.memory_media_file_storage import MemoryMediaFileStorage
 from tests.streamlit.message_mocks import (
     create_dataframe_msg,
     create_script_finished_message,
