+++ lib/tests/streamlit/runtime/scriptrunner/script_runner_test.py
@@ -26,15 +26,16 @@ from tornado.testing import AsyncTestCase
 
 from streamlit import source_util
 from streamlit.elements.exception import _GENERIC_UNCAUGHT_EXCEPTION_TEXT
-from streamlit.runtime import media_file_manager
-from streamlit.runtime.legacy_caching import caching
 from streamlit.proto.ClientState_pb2 import ClientState
 from streamlit.proto.Delta_pb2 import Delta
 from streamlit.proto.Element_pb2 import Element
 from streamlit.proto.ForwardMsg_pb2 import ForwardMsg
 from streamlit.proto.WidgetStates_pb2 import WidgetStates, WidgetState
+from streamlit.runtime import media_file_manager
 from streamlit.runtime.forward_msg_queue import ForwardMsgQueue
+from streamlit.runtime.legacy_caching import caching
 from streamlit.runtime.media_file_manager import MediaFileManager
+from streamlit.runtime.memory_media_file_storage import MemoryMediaFileStorage
 from streamlit.runtime.scriptrunner import (
     ScriptRunner,
     ScriptRunnerEvent,
@@ -49,7 +50,6 @@ from streamlit.runtime.scriptrunner.script_requests import (
 )
 from streamlit.runtime.state.session_state import SessionState
 from streamlit.runtime.uploaded_file_manager import UploadedFileManager
-from streamlit.web.server.memory_media_file_storage import MemoryMediaFileStorage
 from tests import testutil
 
 text_utf = "complete! 👨‍🎤"
